Create WebBrowserRefreshSel1.py
This update will check to see if appropriate drivers are installed and display a dialogue prompt stating that drivers are not present. The options of yes and no will be present and allow user to select to install automatically. If Yes is selected then it will automatically install the drivers, then run the program.